frontend/pages/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.1.0
langchain-google-genai==1.0.0
langchain-community==0.0.13
google-generativeai==0.3.2

# Database
pymongo==4.6.1
motor==3.3.2
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# API and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1